refactor: transfer userRequest.js to services
